Create WebSocket service
